Refactor Admin and Dashboard components for sticky header layout

+++ supabase/migrations/20251106204359_bcbac953-ccc4-4cef-b415-954f5d7947cd.sql
@@ -174,7 +174,7 @@ CREATE OR REPLACE FUNCTION public.create_giveaway_with_codes(
     p_message TEXT,
     p_code_value DECIMAL,
     p_total_codes INTEGER,
-    p_expires_in_hours INTEGER
+    p_expires_in_hours NUMERIC
 )
 RETURNS UUID
 LANGUAGE plpgsql
